fix: use Int64 for notification_record id field to match codebase convention

+++ internal/store/ent/schema/notification_record.go
@@ -16,7 +16,7 @@ type NotificationRecord struct {
 
 func (NotificationRecord) Fields() []ent.Field {
 	return []ent.Field{
-		field.Int("id").Immutable(),
+		field.Int64("id").Immutable(),
 		field.String("uid").NotEmpty(),
 		field.String("channel").NotEmpty(),
 		field.String("template_id").NotEmpty(),
